+++ docx-editor/.changeset/use-print-flow.md
@@ -1,0 +1,5 @@
+---
+'@casualoffice/docs': patch
+---
+
+Internal: extract the print / Export-as-PDF pipeline out of the DocxEditor component into a `usePrintFlow` hook. No behaviour change.
